+++ src/Git_Updater/API/GitHub_API.php
@@ -126,18 +126,11 @@ class GitHub_API extends API implements API_Interface {
 		if ( $this->use_release_asset( $branch_switch ) ) {
 			$release_asset = $this->get_release_asset();
 
-			if ( ! isset( $this->response['release_asset_download'] ) ) {
-				$auth_header = $this->add_auth_header( $this->default_http_get_args, $release_asset );
-				$response    = wp_remote_get( $release_asset, $auth_header );
-				if ( is_wp_error( $response ) ) {
-					return $response;
-				}
-				$response = json_decode( wp_remote_retrieve_body( $response ) );
-				if ( property_exists( $response, 'browser_download_url' ) ) {
-					$this->set_repo_cache( 'release_asset_download', $response->browser_download_url );
-				}
+			if ( empty( $this->response['release_asset_download'] ) ) {
+				$response = $this->api( $release_asset );
+				$this->parse_release_asset_response( $response );
 			}
-			if ( isset( $this->response['release_asset_download'] ) ) {
+			if ( ! empty( $this->response['release_asset_download'] ) ) {
 				return $this->response['release_asset_download'];
 			}
 
@@ -323,6 +316,22 @@ class GitHub_API extends API implements API_Interface {
 		}
 
 		return $branches;
+	}
+
+	/**
+	 * Parse release asset API response.
+	 *
+	 * @param \stdClass $response API response.
+	 *
+	 * @return void
+	 */
+	public function parse_release_asset_response( $response ) {
+		if ( $this->validate_response( $response ) ) {
+			return;
+		}
+		if ( property_exists( $response, 'browser_download_url' ) ) {
+			$this->set_repo_cache( 'release_asset_download', $response->browser_download_url );
+		}
 	}
 
 	/**